pyproject.toml
[project]
name = "dbm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "astropy>=6.0.1",
    "astroquery>=0.4.10",
    "icecream>=2.1.3",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.4",
    "numpy>=1.26.4",
    "pyephem>=9.99",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
]
